+++ packages/cache/src/utils/to-cache.ts
@@ -1,5 +1,5 @@
 import type { Wrapped } from "@rixio/wrapped"
-import { CacheState, createFulfilledCache, createRejectedCache, pendingCache } from "../cache/domain"
+import { CacheState, createFulfilledCache, createRejectedCache, pendingCache } from "../common/domain"
 
 export function toCache<T>(wrapped: Wrapped<T>): CacheState<T> {
 	switch (wrapped.status) {
